Trigger pool move hooks on CG moves
When CG is moved to a pool, the same hooks that are triggered in dataset pool
move are also triggered, with the `obj` parameter being the CG instead of a
dataset.

We also discussed whether we should trigger the generic object_operation hooks
on move operations and decided that we should not.