Add hooks for child entity creation
Currently, begin_fork is the only that can be used to get a notification about a
child creation. It is a low-level hook and may be fired on other cases as well,
such as replica creation. In order to provide a better hooking interface this
patch creates specific hooks for child creation