Add hooks for consistency group deletion
This patch adds a specific hook for deletion of consistency groups. We use a
specific hook and not the generic object deletion in order to pass the
"delete_members" parameter to the hook. Even if "delete_members" is set to True,
we should not trigger individual deletion hooks for each of the consistency
group members.